requirements.txt
psycopg2-binary
pymongo
pandas
matplotlib
tqdm
tabulate
pyyaml